transformed old Droid class into a new
