--- app/src/main/java/com/google/droidjump/DrawableElement.java
@@ -1,69 +1,0 @@
-/*
- * Copyright 2020 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.google.droidjump;
-
-import android.graphics.Bitmap;
-
-public class DrawableElement {
-
-    protected int x;
-    protected int y;
-    protected Bitmap bitmap;
-
-    public DrawableElement(int x, int y, Bitmap bitmap) {
-        this.x = x;
-        this.y = y;
-        this.bitmap = bitmap;
-    }
-
-    public DrawableElement(int x, int y) {
-        this.x = x;
-        this.y = y;
-    }
-
-    public int getX() {
-        return x;
-    }
-
-    public void setX(int x) {
-        this.x = x;
-    }
-
-    public int getY() {
-        return y;
-    }
-
-    public void setY(int y) {
-        this.y = y;
-    }
-
-    public Bitmap getBitmap() {
-        return bitmap;
-    }
-
-    public void setBitmap(Bitmap bitmap) {
-        this.bitmap = bitmap;
-    }
-
-    public int getHeight() {
-        return bitmap.getHeight();
-    }
-
-    public int getWidth() {
-        return bitmap.getWidth();
-    }
-}
