+++ app/src/main/java/com/google/droidjump/GameView.java
@@ -24,6 +24,7 @@ import android.graphics.Color;
 import android.view.MotionEvent;
 import android.view.SurfaceHolder;
 import android.view.SurfaceView;
+import com.google.droidjump.models.*;
 
 public class GameView extends SurfaceView implements Runnable {
 
@@ -56,7 +57,7 @@ public class GameView extends SurfaceView implements Runnable {
         screenMargin = (int) getResources().getDimension(R.dimen.fab_margin);
 
         // Create droid
-        droid = new Droid(screenMargin, screenY - screenMargin, getResources());
+        droid = new Droid(screenMargin, screenY - screenMargin, this.getContext());
     }
 
     @Override
