+++ app/src/main/java/com/google/droidjump/models/Obstacle.java
@@ -16,7 +16,14 @@
 
 package com.google.droidjump.models;
 
+import android.content.res.Resources;
+
 abstract class Obstacle extends GameItem{
+
+    public Obstacle(float x, float y, int pictureID, Resources resources) {
+        super(x, y, pictureID, resources);
+    }
+
     public boolean isVisible(){
         return true;
     }
