+++ app/src/main/java/com/google/droidjump/models/Palm.java
@@ -16,5 +16,10 @@
 
 package com.google.droidjump.models;
 
+import android.content.res.Resources;
+
 public class Palm extends Obstacle {
+    public Palm(float x, float y, int pictureID, Resources resources) {
+        super(x, y, pictureID, resources);
+    }
 }
